suppression de warnings partie 2

+++ PFA - V2/GameScreen.cpp
@@ -229,8 +229,8 @@ void GameScreen::update(void)
 
 		sf::Vector2i _posSelectedArea = Singleton::getInstance().posLeftClickPressed;
 
-		_posSelectedArea.x -= Singleton::getInstance().updatePosLeftClickPressed.x * Chunk::SIZE_OF_CELL;
-		_posSelectedArea.y -= Singleton::getInstance().updatePosLeftClickPressed.y * Chunk::SIZE_OF_CELL;
+		_posSelectedArea.x -= static_cast<int>(Singleton::getInstance().updatePosLeftClickPressed.x * Chunk::SIZE_OF_CELL);
+		_posSelectedArea.y -= static_cast<int>(Singleton::getInstance().updatePosLeftClickPressed.y * Chunk::SIZE_OF_CELL);
 
 		sf::RectangleShape selectionZone(sf::Vector2f(static_cast<float>(mousePos.x - _posSelectedArea.x),
 			static_cast<float>(mousePos.y - _posSelectedArea.y)));
